Move styled components into one file for header
