--- src/App/components/PageHeader/Account/Account.module.css
@@ -1,189 +1,0 @@
-.account_container {
-    display: flex;
-    justify-content: flex-end;
-
-    align-items: center;
-    background: transparent;
-    border-radius: var(--border-radius);
-    gap: var(--gap, 0.5rem);
-}
-
-.ethereum_icon {
-    padding: 0 10px;
-}
-
-.white {
-    white-space: nowrap;
-    font-family: var(--font-deca);
-    font-weight: 300;
-    font-size: var(--header-size);
-    line-height: var(--body-lh);
-    color: var(--text1);
-}
-
-.title_gradient {
-    outline: none;
-    border: none;
-    background: var(--dark2);
-    padding: 7.5px 12px 7.5px 6px;
-    border-radius: 4px;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 4px;
-    cursor: pointer;
-}
-
-.title_gradient:hover,
-.title_gradient:focus-visible {
-    box-shadow: 0px 0px 36px rgba(205, 193, 255, 0.2),
-        0px 0px 21px rgba(205, 193, 255, 0.2),
-        0px 0px 12px rgba(205, 193, 255, 0.2),
-        0px 0px 7px rgba(205, 193, 255, 0.2), 0px 0px 4px var(--accent5),
-        0px 0px 2px rgba(205, 193, 255, 0.2);
-}
-
-.title_gradient p {
-    background: var(--title-gradient);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-}
-
-.more {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    margin-left: 10px;
-}
-.more span {
-    font-size: 18px;
-    margin: 0 1.5px;
-}
-
-.popperContent {
-    width: 100%;
-    min-width: 196px;
-    max-height: 450px;
-    background-color: var(--dark1);
-    box-shadow: rgba(80, 70, 114, 0.01) 0px 0px 1px, rgb(0 0 0 / 4%) 0px 4px 8px,
-        rgb(0 0 0 / 4%) 0px 16px 24px, rgb(0 0 0 / 1%) 0px 24px 32px;
-    padding: 0.5rem;
-    display: flex;
-    flex-direction: column;
-    font-size: 16px;
-
-    color: var(--text2);
-}
-
-.more_row {
-    display: flex;
-    flex: 1 1 0%;
-    flex-direction: row;
-    cursor: pointer;
-    -webkit-box-align: center;
-    align-items: center;
-    padding: 0.5rem;
-    -webkit-box-pack: justify;
-    justify-content: space-between;
-    color: rgb(195, 197, 203);
-}
-
-.wallet_display {
-    position: relative;
-}
-
-.wallet_wrapper,
-.wallet_wrapper_active {
-    transition: all 0.6s cubic-bezier(0.68, -0.55, 0.265, 1.55);
-    box-shadow: 0px 45px 30px rgba(0, 0, 0, 0.5);
-    background: var(--dark2);
-    position: absolute;
-    top: 50px;
-    width: 347px;
-    height: 280px;
-    right: -40px;
-    z-index: 9999;
-    padding: 8px 1rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    gap: 1rem;
-    border-radius: var(--border-radius);
-}
-
-.wallet_wrapper {
-    bottom: 0px;
-    opacity: 0;
-    pointer-events: none;
-}
-
-.wallet_wrapper_active {
-    opacity: 1;
-    bottom: 85px;
-    pointer-events: auto;
-}
-
-.block_number_div {
-    margin-right: 10px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 0.25rem;
-}
-
-.block_number_div span {
-    cursor: default !important;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text1);
-    color: var(--text1);
-}
-
-/* page block sign */
-.page_block_sign {
-    cursor: default !important;
-    width: 10px;
-    height: 10px;
-    border-radius: 50%;
-    background: linear-gradient(45deg, transparent, transparent 40%, yellow),
-        linear-gradient(#262626, #141414);
-    animation: rotate 9.5s linear infinite;
-    color: var(--text1);
-}
-
-.page_block_sign:before,
-.page_block_sign:after {
-    content: ' ';
-    position: absolute;
-    background: linear-gradient(#161616, #262626);
-    border-radius: inherit;
-}
-.page_block_sign:before {
-    background: linear-gradient(
-        45deg,
-        transparent,
-        transparent 40%,
-        rgb(40, 2, 2)
-    );
-    filter: blur(32px);
-}
-
-.wallet_name {
-    min-width: 80px;
-}
-
-@keyframes rotate {
-    100% {
-        transform: rotate(360deg);
-        filter: hue-rotate(360deg);
-    }
-}
-
-@media (max-width: 600px) {
-    .wallet_name {
-        display: none;
-    }
-}
