+++ src/styled/Components/Header.ts
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom';
 import styled, { css } from 'styled-components';
 import { motion } from 'framer-motion';
 
+import { FlexContainer, GridContainer } from '../Common';
 const fixedStyles = `   
 position: fixed;
 top: 0;
@@ -90,27 +91,6 @@ export const TradeNowDiv = styled.div`
     padding: 0 1rem;
 `;
 
-// Define the styles for AccountDiv
-export const AccountDiv = styled.div`
-    display: flex;
-    align-items: center;
-    gap: 8px;
-`;
-
-// Define the styles for BranchNameDiv
-export const BranchNameDiv = styled.div`
-    font-size: 12px;
-    color: orange;
-`;
-
-// Define the styles for BranchDiv
-export const BranchDiv = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 4px;
-`;
-
 interface PrimaryNavigationProps {
     dataVisible: boolean;
 }
@@ -118,11 +98,10 @@ export const PrimaryNavigation = styled.nav<PrimaryNavigationProps>`
     display: flex;
     flex-direction: row;
     align-items: center;
-    gap: var(--gap, 2rem);
+    gap: 2rem;
 
     @media (max-width: 800px) {
         display: none;
-        --gap: 2em;
         position: fixed;
         z-index: 1000;
         inset: 0 0 0 30%;
@@ -236,5 +215,426 @@ export const AuthenticateButton = styled.button<AuthenticateButtonProps>`
             0px 0px 12px rgba(205, 193, 255, 0.2),
             0px 0px 7px rgba(205, 193, 255, 0.2), 0px 0px 4px var(--accent5),
             0px 0px 2px rgba(205, 193, 255, 0.2);
+    }
+`;
+
+// WalletDropdown
+
+export const WalletDisplay = styled(FlexContainer)`
+    p {
+        font-size: var(--body-size);
+        line-height: var(--body-lh);
+        color: var(--text2);
+    }
+`;
+
+export const NameDisplay = styled(FlexContainer)`
+    h2 {
+        font-size: var(--header2-size);
+        line-height: var(--header2-lh);
+        color: var(--text1);
+        font-weight: 100;
+    }
+`;
+
+export const CopyButton = styled.button`
+    cursor: pointer;
+    background: transparent;
+    border: none;
+`;
+
+export const TokenContainer = styled.section`
+    font-size: var(--header2-size);
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    padding: 4px 10px;
+
+    &:not(:last-of-type) {
+        border-bottom: 1px solid var(--dark1);
+    }
+`;
+
+export const LogoName = styled(FlexContainer)`
+    img {
+        width: 25px;
+        height: 25px;
+    }
+`;
+
+export const TokenAmount = styled(FlexContainer)`
+    h3 {
+        color: var(--text1);
+        text-align: right;
+    }
+
+    h3,
+    h6 {
+        text-align: right;
+    }
+`;
+
+export const NameDisplayContainer = styled(FlexContainer)`
+    margin: 0 auto;
+    overflow: visible;
+    .image {
+        width: 50px;
+        height: 50px;
+        border-radius: 50%;
+        background: var(--title-gradient);
+    }
+`;
+
+export const ActionsContainer = styled(GridContainer)`
+    bottom: 0;
+    z-index: 9;
+
+    a:hover,
+    a:focus-visible {
+        color: var(--text-grey-white);
+    }
+`;
+
+export const WalletContent = styled.section`
+    background: var(--dark3);
+    padding: 8px 1rem;
+    border-radius: var(--border-radius);
+    display: flex;
+    flex-direction: column;
+    height: 200px;
+    overflow-y: hidden;
+    &::-webkit-scrollbar {
+        display: none;
+    }
+`;
+
+export const WalletWrapper = styled(FlexContainer)`
+    transition: all 0.6s cubic-bezier(0.68, -0.55, 0.265, 1.55);
+    box-shadow: 0px 45px 30px rgba(0, 0, 0, 0.5);
+    background: var(--dark2);
+    position: absolute;
+    top: 50px;
+    width: 347px;
+    height: 280px;
+    right: -40px;
+    z-index: 9999;
+    padding: 8px 1rem;
+
+    opacity: 1;
+    bottom: 85px;
+    pointer-events: auto;
+`;
+
+// Account
+
+export const TitleGradientButton = styled.button`
+    outline: none;
+    border: none;
+    background: var(--dark2);
+    padding: 7.5px 12px 7.5px 6px;
+    border-radius: 4px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 4px;
+    cursor: pointer;
+
+    &:hover,
+    &:focus-visible {
+        box-shadow: 0px 0px 36px rgba(205, 193, 255, 0.2),
+            0px 0px 21px rgba(205, 193, 255, 0.2),
+            0px 0px 12px rgba(205, 193, 255, 0.2),
+            0px 0px 7px rgba(205, 193, 255, 0.2), 0px 0px 4px var(--accent5),
+            0px 0px 2px rgba(205, 193, 255, 0.2);
+    }
+
+    p {
+        background: var(--title-gradient);
+        -webkit-background-clip: text;
+        -webkit-text-fill-color: transparent;
+        background-clip: text;
+        font-size: var(--body-size);
+        line-height: var(--body-lh);
+    }
+`;
+
+export const WalletName = styled.p`
+    min-width: 80px;
+
+    @media (max-width: 600px) {
+        display: none;
+    }
+`;
+
+// ExchangeBalanceDropdown
+
+export const StyledExchangeBalanceDropdown = styled.div`
+    position: absolute;
+    top: 68px;
+    right: 8px;
+    width: 450px;
+    height: 510px;
+    border: none;
+    overflow: hidden;
+    transition: all var(--animation-speed) ease;
+    z-index: 999;
+    background: var(--dark1);
+    border-radius: var(--border-radius);
+    box-shadow: 4px 4px 12px rgba(0, 0, 0, 0.25);
+    text-align: start;
+
+    @media only screen and (max-width: 600px) {
+        width: auto;
+        height: auto;
+        background: transparent;
+        border-radius: 4px;
+    }
+`;
+
+// NavbarDropdownMenu
+
+export const NavbarDropdown = styled.div`
+    position: absolute;
+    top: 60px;
+    width: 240px;
+    height: 376px;
+    transform: translateX(-45%);
+    border: none;
+    padding: 1rem;
+    overflow: hidden;
+    transition: all var(--animation-speed) ease;
+
+    background: var(--dark2);
+    border-radius: var(--border-radius);
+    box-shadow: 4px 4px 12px rgba(0, 0, 0, 0.25);
+    z-index: 999;
+`;
+
+export const Menu = styled(motion.div)`
+    width: 100%;
+    z-index: 999;
+    outline: none;
+    border: none;
+    background: transparent;
+`;
+
+export const MenuItem = styled.div`
+    width: 100%;
+    height: 50px;
+    display: flex;
+    align-items: center;
+    border-radius: var(--border-radius);
+    padding: 0.5rem;
+    cursor: pointer;
+    color: var(--text1);
+    font-family: var(--font-deca);
+    font-weight: 300;
+    font-size: var(--header2-size);
+    line-height: var(--header2-lh);
+    outline: none;
+    border: none;
+    background: transparent;
+
+    transition: all var(--animation-speed) ease-out;
+    background-position: 1% 50%;
+    background-size: 400% 300%;
+    text-decoration: none;
+
+    &:hover,
+    &:focus-visible {
+        filter: drop-shadow(0 0 5px var(--text1));
+    }
+`;
+
+export const NavbarLogoutContainer = styled(FlexContainer)`
+    margin-top: 8px;
+
+    &:focus-visible {
+        outline: none;
+    }
+`;
+
+export const ConnectButton = styled.button`
+    cursor: pointer;
+    width: 100%;
+    font-size: var(--header2-size);
+    line-height: var(--header2-lh);
+    color: var(--text1);
+    border: none;
+    outline: none;
+    background: var(--accent1);
+    padding: 6px 8px;
+    gap: 4px;
+    border-radius: var(--border-radius);
+
+    &:hover,
+    &:focus-visible {
+        color: var(--accent1);
+        background: var(--dark2);
+        border: solid 1px var(--accent1);
+    }
+`;
+
+export const IconButton = styled.div`
+    margin-right: 0.5rem;
+    background: var(--blur-bg);
+    border-radius: 50%;
+    padding: 3px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-transform: capitalize;
+
+    &:hover {
+        filter: none;
+    }
+`;
+
+export const NonTopLevelContainer = styled.div`
+    &:hover,
+    &:focus-visible {
+        background: var(--dark2);
+    }
+`;
+
+export const GoBackStyle = styled.div`
+    &:hover,
+    &:focus-visible {
+        background: var(--dark2);
+    }
+
+    .icon_button {
+        background: var(--title-gradient);
+
+        svg {
+            color: black;
+            transition: all var(--animation-speed) ease;
+        }
+    }
+`;
+
+export const IconRight = styled.span`
+    margin-left: auto;
+    svg {
+        color: var(--text1);
+    }
+`;
+
+export const ButtonContainer = styled.div`
+    width: 100%;
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 0 10px;
+`;
+
+// NavItem
+
+const buttonSize = 'calc(var(--nav-size) * 0.5)';
+
+export const NavItemButton = styled.button`
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    outline: none;
+    border: none;
+    background: transparent;
+
+    &:focus-visible .icon_button {
+        filter: brightness(1.2);
+        background: var(--dark2);
+    }
+`;
+
+interface IconButtonProps {
+    square: boolean | undefined;
+}
+
+export const NavItemIconButton = styled.div<IconButtonProps>`
+    height: ${buttonSize};
+    background-color: var(--dark2);
+    border-radius: ${(props) => (props.square ? '4px' : '50%')};
+    padding: 5px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: all var(--animation-speed) ease-in-out;
+    cursor: pointer;
+
+    &:hover,
+    &:focus-visible {
+        filter: brightness(1.2);
+        background: var(--dark2);
+    }
+
+    svg {
+        fill: var(--text1);
+        width: 20px;
+        height: 20px;
+    }
+`;
+
+// NetworkSelector
+
+export const DropdownMenuContainer = styled.div`
+    z-index: 99999;
+    background-color: var(--dark2);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 0 4px;
+    border-radius: 4px;
+    gap: 4px;
+    height: 31px;
+`;
+
+export const MenuContent = styled.ul`
+    width: 200px;
+    padding: 12px;
+    overflow: hidden;
+    z-index: 2;
+    background: var(--dark2);
+    border-radius: var(--border-radius);
+    z-index: 999;
+
+    ul,
+    li {
+        text-decoration: none;
+        list-style-type: none;
+    }
+`;
+
+export const NetworkItem = styled(motion.li)`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    transition: background 500ms;
+    padding: 8px 0;
+    cursor: pointer;
+    text-decoration: none;
+
+    &:hover {
+        backdrop-filter: blur(16.5px);
+        -webkit-backdrop-filter: blur(16.5px);
+        border-radius: 4px;
+    }
+`;
+
+export const ChainNameStatus = styled.div`
+    padding: 0.6rem 0;
+    font-size: var(--header2-size);
+    width: 100%;
+    color: var(--text1);
+    z-index: 9999;
+
+    &:hover,
+    &:focus-visible {
+        filter: drop-shadow(0 0 5px var(--text1));
+    }
+
+    img {
+        margin-right: 0.5em;
+        vertical-align: middle;
     }
 `;
